Fix awful random scalar generating code
This code is nasty:

- every time through the loop requires a mutex acquire and release
rather than holding the mutex around the loop.

- the comment about the limit is confusing af: eventually I figured out
that the "limit" equals 15L, and that this code is basically trying to
generate a value in (0, 15L) then mod L the number to get an unbiased
value in (0, L).

- However, the code is broken in two ways:
  - it can return 0 because (nL % L == 0) for integer n > 0
  - the while condition is broken af so that the while loop can *never*
    repeat, and thus the entire function is just always returning
    (random % L) which thus has the slight bias.

This commit fixes all these issues by moving the mutex outside the loop
and borrowing libsodium's approach for the generation: generate a random
value, mask off the 3 most significant bits, then repeat until this
yields a value in (0, L), which happens slightly more than 1/2 the time.

This is a bit slower than the 15L approach, but much simpler and
generating random scalars is not a performance bottleneck.